requirements.txt
imageio==2.22.4
joblib==1.2.0
kiwisolver==1.4.4
kornia==0.6.8
numpy==1.23.5
opencv-python==4.6.0.66
pandas==1.5.2
Pillow==9.3.0
pyparsing==3.0.9
PyYAML==6.0
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1